requirements.txt
discord.py
requests
aiohttp
pytesseract
python-dotenv
flask